Test #2. Attempt to fix mistakes in last print.
Now its certainly loose enough, too loose.  Try for 0.5mm gap next time.
Also, try keeping the channel the same size as the pin and rely on the gap
to make it easy to move the pin around.  I'm a little concerned that the
cylinder is pressing into the top of the lid and keeping it from sitting tight.  I'm going to try making the lid a bit longer.