Test #8.
The pin inside the lid had 100 facets instead of 4 which made the last one very
tight.  The last one was also super small, so this one is significantly
enlarged.  Also, the pin was in the wrong place and did not allow the lid to
sit fully down on the handle, so it was adjusted to sit properly.  Since the
pixel diameter was enlarged, lots of other variables had to change including
the number of rows that are "empty" at the top and bottom.

Update after printing:  This is almost perfect!  The size is a bit enlarged,
but it looks nice.  The channels are nice and large, easy to see and yet they
are still tight.  I do need to make the maze a bit simpler.  The main problem
is that the bump is way too large and it its really hard to snap it into place.
The location vertically of the pin is perfect.