Test #10 is a success! I just need to enlarge the base to lid gap a smidge, trying 0.70 instead of 0.65. This should both help it move more smoothly and make the bump easier to get over. Kristy said the maze is a bit too easy to solve now. I'm also working on enlarging the maze so the puzzle is a bit taller and more in proportion to the original Thingiverse maze.
I'd also like to set it up so the kids can easily fill in a 12x12 grid
and then use some software to create a 36x36 grid with proper spacing
to create a maze puzzle.  Then they don't have to worry about the spacing
and I can automate the activity.  To go even further, if I could scan in
a picture of the maze, maybe I could create a maze from that.